+++ K8sUtils/public/Get-PodStatus.ps1
@@ -53,7 +53,7 @@ $podStatuses = @{}
 $createdTempFile = !$OutputFile
 
 if ($IsJob) {
-    $phases = "Succeeded","Running" # running for timeout
+    $phases = ,"Succeeded"
     $prefix = "preHook job pod"
 } else {
     $phases = ,"Running"
@@ -62,15 +62,19 @@ if ($IsJob) {
 
 # are all containers in the pod ready?
 function podReady($containerStatuses) {
+
+    $readyContainers = @()
+    Write-Warning "ContainerStatuses: $($containerStatuses | ConvertTo-Json -Depth 10 -EnumsAsStrings)"
     if ($IsJob) {
-        Write-Verbose "Checking containerStatuses for job" # that should have exited
-        Write-Verbose ($containerStatuses | ConvertTo-Json -Depth 10 -EnumsAsStrings)
-        # return $true
-        return !($containerStatuses | Where-Object { (Get-Member -InputObject $_.state -Name 'terminated') -and  $_.state.terminated.reason -ne 'Completed'})
+        # for a job, all containers need to be complete
+        $readyContainers += $containerStatuses | Where-Object { (Get-Member -InputObject $_.state -Name 'terminated') -and $_.state.terminated.reason -eq 'Completed'}
     } else {
-        Write-Verbose "Checking containerStatuses for NON job" # that should be running
-        return !($containerStatuses | Where-Object ready -ne $true) # all containers ready
+        # for deploys all should be running
+        $readyContainers += $containerStatuses | Where-Object ready -eq $true
     }
+    $podIsReady = $readyContainers.Count -eq $containerStatuses.Count
+    Write-Verbose "Checking containerStatuses for $($IsJob ? 'job' : 'NON job'). PodReady = $podIsReady"
+    return $podIsReady
 }
 
 if (!$OutputFile) {
@@ -135,13 +139,14 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
             if (!$runningPods[$pod.metadata.name] -and
                 (podReady $pod.status.containerStatuses)) {
 
-                "  $prefix $($pod.metadata.name) has all containers ready or completed or timed out`n" | Tee-Object $OutputFile -Append | Write-MyHost
+                $status = $IsJob ? [Status]::Completed : [Status]::Running
+                "  $prefix $($pod.metadata.name) has all containers ready or completed. Status is $status`n" | Tee-Object $OutputFile -Append | Write-MyHost
                 $runningCount += 1
                 $runningPods[$pod.metadata.name] = $true
-                $podStatuses[$pod.metadata.name].Status = [Status]::Running
-                $podStatuses[$pod.metadata.name].ContainerStatuses = @($pod.status.containerStatuses | ForEach-Object { [ContainerStatus]::new($_.name, [Status]::Running) })
+                $podStatuses[$pod.metadata.name].Status = $status
+                $podStatuses[$pod.metadata.name].ContainerStatuses = @($pod.status.containerStatuses | ForEach-Object { [ContainerStatus]::new($_.name, $status) })
                 if ($HasInit) {
-                    $podStatuses[$pod.metadata.name].InitContainerStatuses = @($pod.status.initContainerStatuses | ForEach-Object { [ContainerStatus]::new($_.name, [Status]::Running) })
+                    $podStatuses[$pod.metadata.name].InitContainerStatuses = @($pod.status.initContainerStatuses | ForEach-Object { [ContainerStatus]::new($_.name, $status) })
                 }
 
                 # write final events and logs for this pod
@@ -150,10 +155,11 @@ while ($runningCount -lt $ReplicaCount -and !$timedOut)
                 Write-PodLog -Prefix $prefix -PodName $pod.metadata.name -Namespace $Namespace -LogLevel ok -HasInit:$HasInit
                 continue
             } else {
-                Write-Verbose "Pod $($pod.metadata.name) is ready, but pod containerStatuses are: $($pod.status.containerStatuses  | out-string)"
+                Write-Verbose "Pod $($pod.metadata.name) is ready, but pod containerStatuses are: $($pod.status.containerStatuses | out-string)"
             }
         }
 
+        # check for any errors since not ready yet
         $lastEventTime = Get-Date
         Write-Verbose "kubectl get events --namespace $Namespace --field-selector `"involvedObject.name=$($pod.metadata.name)`" -o json"
         $events = kubectl get events --namespace $Namespace --field-selector "involvedObject.name=$($pod.metadata.name)" -o json | ConvertFrom-Json