+++ Tools/TestHelpers.ps1
@@ -18,7 +18,7 @@ function Test-Deploy( $deploy, $running = $true, $podCount = 1, $rollbackStatus 
 function Test-Pod( $podStatus, $status, $containerStatus, $reason, $nameLike, $containerName ) {
     $podStatus.Status | Should -Be $status
     $podStatus.PodName | Should -BeLike $nameLike
-    if ($reason -ne "Possible timeout") {
+    if ($reason -ne "Possible timeout" -and $status -ne 'Timeout') { # timeouts don't save container statuses
         $podStatus.ContainerStatuses | Should -Not -BeNullOrEmpty
         $podStatus.ContainerStatuses.Count | Should -Be 1
         $podStatus.ContainerStatuses[0].ContainerName | Should -Be $containerName
@@ -29,7 +29,7 @@ function Test-Pod( $podStatus, $status, $containerStatus, $reason, $nameLike, $c
     }
 }
 
-function Test-PreHook( $podStatus,  $status = 'Running', $containerStatus = 'Running', $reason = $null) {
+function Test-PreHook( $podStatus, $status = 'Completed', $containerStatus = 'Completed', $reason = $null) {
     Test-Pod $podStatus $status $containerStatus $reason 'test-prehook-*' 'pre-install-upgrade-job'
 }